Implement subscription using Context API and useReducer
